feat: add solutions to lc problem: No.0363
No.0363.Max Sum of Rectangle No Larger Than K
