+++ solution/0300-0399/0363.Max Sum of Rectangle No Larger Than K/Solution.ts
@@ -1,198 +1,3 @@
-# [363. 矩形区域不超过 K 的最大数值和](https://leetcode.cn/problems/max-sum-of-rectangle-no-larger-than-k)
-
-[English Version](/solution/0300-0399/0363.Max%20Sum%20of%20Rectangle%20No%20Larger%20Than%20K/README_EN.md)
-
-## 题目描述
-
-<!-- 这里写题目描述 -->
-
-<p>给你一个 <code>m x n</code> 的矩阵 <code>matrix</code> 和一个整数 <code>k</code> ，找出并返回矩阵内部矩形区域的不超过 <code>k</code> 的最大数值和。</p>
-
-<p>题目数据保证总会存在一个数值和不超过 <code>k</code> 的矩形区域。</p>
-
-<p> </p>
-
-<p><strong>示例 1：</strong></p>
-<img alt="" src="https://fastly.jsdelivr.net/gh/doocs/leetcode@main/solution/0300-0399/0363.Max%20Sum%20of%20Rectangle%20No%20Larger%20Than%20K/images/sum-grid.jpg" style="width: 255px; height: 176px;" />
-<pre>
-<strong>输入：</strong>matrix = [[1,0,1],[0,-2,3]], k = 2
-<strong>输出：</strong>2
-<strong>解释：</strong>蓝色边框圈出来的矩形区域 <code>[[0, 1], [-2, 3]]</code> 的数值和是 2，且 2 是不超过 k 的最大数字（k = 2）。
-</pre>
-
-<p><strong>示例 2：</strong></p>
-
-<pre>
-<strong>输入：</strong>matrix = [[2,2,-1]], k = 3
-<strong>输出：</strong>3
-</pre>
-
-<p> </p>
-
-<p><strong>提示：</strong></p>
-
-<ul>
-	<li><code>m == matrix.length</code></li>
-	<li><code>n == matrix[i].length</code></li>
-	<li><code>1 <= m, n <= 100</code></li>
-	<li><code>-100 <= matrix[i][j] <= 100</code></li>
-	<li><code>-10<sup>5</sup> <= k <= 10<sup>5</sup></code></li>
-</ul>
-
-<p> </p>
-
-<p><strong>进阶：</strong>如果行数远大于列数，该如何设计解决方案？</p>
-
-## 解法
-
-<!-- 这里可写通用的实现逻辑 -->
-
-**方法一：枚举边界 + 有序集合**
-
-我们可以枚举矩形的上下边界 $i$ 和 $j$，然后计算出该边界内每列的元素和，记录在数组 $nums$ 中，问题转化为如何在数组 $nums$ 中寻找不超过 $k$ 的最大子数组和。
-
-我们可以使用有序集合来快速寻找小于等于 $x$ 的最大值，从而得到最大子数组和不超过 $k$ 的子数组。
-
-时间复杂度 $O(m^2 \times n \times \log n)$，空间复杂度 $O(n)$。
-
-<!-- tabs:start -->
-
-### **Python3**
-
-<!-- 这里可写当前语言的特殊实现逻辑 -->
-
-```python
-from sortedcontainers import SortedSet
-
-
-class Solution:
-    def maxSumSubmatrix(self, matrix: List[List[int]], k: int) -> int:
-        m, n = len(matrix), len(matrix[0])
-        ans = -inf
-        for i in range(m):
-            nums = [0] * n
-            for j in range(i, m):
-                for h in range(n):
-                    nums[h] += matrix[j][h]
-                s = 0
-                ts = SortedSet([0])
-                for x in nums:
-                    s += x
-                    p = ts.bisect_left(s - k)
-                    if p != len(ts):
-                        ans = max(ans, s - ts[p])
-                    ts.add(s)
-        return ans
-```
-
-### **Java**
-
-<!-- 这里可写当前语言的特殊实现逻辑 -->
-
-```java
-class Solution {
-    public int maxSumSubmatrix(int[][] matrix, int k) {
-        int m = matrix.length;
-        int n = matrix[0].length;
-        final int inf = 1 << 30;
-        int ans = -inf;
-        for (int i = 0; i < m; ++i) {
-            int[] nums = new int[n];
-            for (int j = i; j < m; ++j) {
-                for (int h = 0; h < n; ++h) {
-                    nums[h] += matrix[j][h];
-                }
-                int s = 0;
-                TreeSet<Integer> ts = new TreeSet<>();
-                ts.add(0);
-                for (int x : nums) {
-                    s += x;
-                    Integer y = ts.ceiling(s - k);
-                    if (y != null) {
-                        ans = Math.max(ans, s - y);
-                    }
-                    ts.add(s);
-                }
-            }
-        }
-        return ans;
-    }
-}
-```
-
-### **C++**
-
-```cpp
-class Solution {
-public:
-    int maxSumSubmatrix(vector<vector<int>>& matrix, int k) {
-        int m = matrix.size(), n = matrix[0].size();
-        const int inf = 1 << 30;
-        int ans = -inf;
-        for (int i = 0; i < m; ++i) {
-            vector<int> nums(n);
-            for (int j = i; j < m; ++j) {
-                for (int h = 0; h < n; ++h) {
-                    nums[h] += matrix[j][h];
-                }
-                set<int> ts;
-                int s = 0;
-                ts.insert(0);
-                for (int x : nums) {
-                    s += x;
-                    auto it = ts.lower_bound(s - k);
-                    if (it != ts.end()) {
-                        ans = max(ans, s - *it);
-                    }
-                    ts.insert(s);
-                }
-            }
-        }
-        return ans;
-    }
-};
-```
-
-### **Go**
-
-```go
-func maxSumSubmatrix(matrix [][]int, k int) int {
-	m, n := len(matrix), len(matrix[0])
-	const inf = 1 << 30
-	ans := -inf
-	for i := 0; i < m; i++ {
-		nums := make([]int, n)
-		for j := i; j < m; j++ {
-			for h := 0; h < n; h++ {
-				nums[h] += matrix[j][h]
-			}
-			s := 0
-			rbt := redblacktree.NewWithIntComparator()
-			rbt.Put(0, nil)
-			for _, x := range nums {
-				s += x
-				if y, ok := rbt.Ceiling(s - k); ok {
-					ans = max(ans, s-y.Key.(int))
-				}
-				rbt.Put(s, nil)
-			}
-		}
-
-	}
-	return ans
-}
-
-func max(a, b int) int {
-	if a > b {
-		return a
-	}
-	return b
-}
-```
-
-### **TypeScript**
-
-```ts
 function maxSumSubmatrix(matrix: number[][], k: number): number {
     const m = matrix.length;
     const n = matrix[0].length;
@@ -858,12 +663,3 @@ class TreeMultiSet<T = number> {
         return undefined;
     }
 }
-```
-
-### **...**
-
-```
-
-```
-
-<!-- tabs:end -->
